requirements.txt
# Streamlit App Core
streamlit==1.32.0
pandas==2.1.0
requests==2.31.0
python-dotenv==1.0.0

# ESPN fantasy league integration
espn_api==0.45.0

# OpenAI integration for AI trade verdicts
openai==0.27.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
html5lib==1.1

# Optional (safe fallback for legacy environments)
dataclasses; python_version < "3.7"
